Fix printing regular tensors inside functorch transforms
Fixes https://github.com/pytorch/functorch/issues/1026

We need to disable functorch's stack-based dispatching mechanism inside
the tensor printing. Otherwise, all operations that clean up the data of
the Tensor for printing dispatch through the entire functorch stack and
causes problems.

Disabling stack-based dispatching and printing a functorch wrapped
tensor is not a problem; we're still able to get the attributes on the
wrapped tensor that we want.

Test Plan:
- new test

ghstack-source-id: 35c450721877958585438f4157f250aa12a5acc3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85556