[Profiler][Minor] Group and consolidate stub APIs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85510

There is a concept in profiler of a stub that wraps a profiling API. It was introduced for CUDA profiling before Kineto, and ITT has adopted it to call into VTune APIs. However for the most part we don't really interact with them when developing the PyTorch profiler.

Thus it makes sense to unify the fallback registration mechanism and create a subfolder to free up real estate in the top level `torch/csrc/profiler` directory.
ghstack-source-id: 168347018

Differential Revision: [D39108647](https://our.internmc.facebook.com/intern/diff/D39108647/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D39108647/)!