TestModule: Don't assume sample inputs version counter is zero
The intention of this assert is to check the input tensor's version
counter has increased, indicating it was mutated by `m_inplace`.
However, the cloning step to create `input_arg_clone` restarts the
version counter to zero, so this test may fail if the sample input
was ever mutated during its creation.

ghstack-source-id: 9e3d240a979e812d9cecabd8cdb194f9e2fcff82
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85734